requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
yfinance==0.2.10
SQLAlchemy==2.0.2
pydantic==1.10.4
Jinja2==3.1.2
python-dotenv==0.21.1
psycopg2==2.9.5